Add support for the Orleans binary serializer in MongoDB Grain state
