+++ Orleans.Providers.MongoDB/StorageProviders/BaseJSONStorageProvider.cs
@@ -6,6 +6,12 @@ using Newtonsoft.Json;
 using Orleans.Runtime;
 using Orleans.Serialization;
 using Orleans.Storage;
+using Newtonsoft.Json.Linq;
+using MongoDB.Bson;
+using MongoDB.Bson.Serialization;
+using System.Text;
+using Orleans.Providers.MongoDB.StorageProviders.Serializing;
+using System.IO;
 
 namespace Orleans.Providers.MongoDB.StorageProviders
 {
@@ -15,12 +21,14 @@ namespace Orleans.Providers.MongoDB.StorageProviders
     public abstract class BaseJSONStorageProvider : IStorageProvider
     {
         private JsonSerializerSettings serializerSettings;
-
+        
         /// <summary>
         ///     Data manager instance
         /// </summary>
         /// <remarks>The data manager is responsible for reading and writing JSON strings.</remarks>
         protected IJSONStateDataManager DataManager { get; set; }
+
+        private SerializationManager serializationManager;
 
         /// <summary>
         ///     Logger object
@@ -33,6 +41,11 @@ namespace Orleans.Providers.MongoDB.StorageProviders
         public string Name { get; protected set; }
 
         /// <summary>
+        ///     Use JSON or Binary serialization
+        /// </summary>
+        public bool UseJsonFormat { get; set; }
+
+        /// <summary>
         ///     Initializes the storage provider.
         /// </summary>
         /// <param name="name">The name of this provider instance.</param>
@@ -43,11 +56,12 @@ namespace Orleans.Providers.MongoDB.StorageProviders
         {
             Log = providerRuntime.GetLogger(GetType().FullName);
 
-            var serializationManager = providerRuntime.ServiceProvider.GetRequiredService<SerializationManager>();
+            serializationManager = providerRuntime.ServiceProvider.GetRequiredService<SerializationManager>();
             serializerSettings = OrleansJsonSerializer.UpdateSerializerSettings(
                 OrleansJsonSerializer.GetDefaultSerializerSettings(serializationManager, providerRuntime.GrainFactory),
                 config);
 
+            UseJsonFormat = config.GetBoolProperty("UseJsonFormat", true);
             serializerSettings.Converters.Add(new GrainReferenceConverter(providerRuntime.GrainFactory));
             return Task.CompletedTask;
         }
@@ -132,9 +146,22 @@ namespace Orleans.Providers.MongoDB.StorageProviders
         ///     http://msdn.microsoft.com/en-us/library/system.web.script.serialization.javascriptserializer.aspx
         ///     for more on the JSON serializer.
         /// </remarks>
-        protected string ConvertToStorageFormat(IGrainState grainState)
+        protected BsonDocument ConvertToStorageFormat(IGrainState grainState)
         {
-            return JsonConvert.SerializeObject(grainState.State, serializerSettings);
+            if (UseJsonFormat)
+            {
+                var json = JsonConvert.SerializeObject(grainState.State, serializerSettings);
+                json = ReverseInvalidValues(json);
+                return BsonSerializer.Deserialize<BsonDocument>(json);
+            }
+            else
+            {
+                var bindata = serializationManager.SerializeToByteArray(grainState.State);
+                return new BsonDocument
+                {
+                    ["statedata"] = bindata,
+                };
+            }
         }
 
         /// <summary>
@@ -142,9 +169,120 @@ namespace Orleans.Providers.MongoDB.StorageProviders
         /// </summary>
         /// <param name="grainState">Grain state to be populated for storage.</param>
         /// <param name="entityData">JSON storage format representaiton of the grain state.</param>
-        protected void ConvertFromStorageFormat(IGrainState grainState, string entityData)
+        protected void ConvertFromStorageFormat(IGrainState grainState, BsonDocument entityData)
         {
-            JsonConvert.PopulateObject(entityData, grainState.State, serializerSettings);
+            if(UseJsonFormat)
+            {
+                var json = entityData.ToJson();
+                json = ReverseInvalidValues(json);
+                JsonConvert.PopulateObject(json, grainState.State, serializerSettings);
+            }
+            else
+            {
+                grainState.State = serializationManager.DeserializeFromByteArray<object>((byte[])entityData["statedata"]);
+            }
+        }
+
+        /// <summary>
+        ///     NewtonSoft generates a $type & $id which is incompatible with Mongo. Replacing $ with __
+        /// </summary>
+        /// <param name="entityData"></param>
+        /// <returns></returns>
+        private string ReverseInvalidValues(string entityData)
+        {
+            var sb = new StringBuilder();
+            var writer = new CustomJsonWriter(new StringWriter(sb));
+            var reader = new JsonTextReader(new StringReader(entityData));
+
+            while (reader.Read())
+                if (reader.Value != null)
+                {
+                    switch (reader.TokenType)
+                    {
+                        case JsonToken.PropertyName:
+
+                            if (reader.Value.ToString() == "$type")
+                                writer.WritePropertyName("__type");
+                            else if (reader.Value.ToString() == "__type")
+                                writer.WritePropertyName("$type");
+                            else if (reader.Value.ToString() == "$id")
+                                writer.WritePropertyName("__id");
+                            else if (reader.Value.ToString() == "__id")
+                                writer.WritePropertyName("$id");
+                            else if (reader.Value.ToString() == "$type")
+                                writer.WritePropertyName("__type");
+                            else if (reader.Value.ToString() == "__type")
+                                writer.WritePropertyName("$type");
+                            else if (reader.Value.ToString() == "$values")
+                                writer.WritePropertyName("__values");
+                            else if (reader.Value.ToString() == "__values")
+                                writer.WritePropertyName("$values");
+                            else
+                                writer.WritePropertyName(reader.Value.ToString());
+
+                            break;
+                        case JsonToken.None:
+                            break;
+                        case JsonToken.StartConstructor:
+                            writer.WriteStartConstructor(reader.Value.ToString());
+                            break;
+                        case JsonToken.Comment:
+                            writer.WriteComment(reader.Value.ToString());
+                            break;
+                        case JsonToken.Raw:
+                            writer.WriteRaw(reader.Value.ToString());
+                            break;
+                        case JsonToken.Integer:
+                        case JsonToken.Float:
+                        case JsonToken.String:
+                        case JsonToken.Boolean:
+                        case JsonToken.Date:
+                        case JsonToken.Bytes:
+                            writer.WriteValue(reader.Value);
+                            break;
+                        case JsonToken.Null:
+                            writer.WriteNull();
+                            break;
+                        case JsonToken.Undefined:
+                            writer.WriteUndefined();
+                            break;
+                        default:
+                            break;
+                    }
+                }
+                else
+                {
+                    switch (reader.TokenType)
+                    {
+                        case JsonToken.None:
+                            break;
+                        case JsonToken.StartObject:
+                            writer.WriteStartObject();
+                            break;
+                        case JsonToken.StartArray:
+                            writer.WriteStartArray();
+                            break;
+                        case JsonToken.Null:
+                            writer.WriteNull();
+                            break;
+                        case JsonToken.Undefined:
+                            writer.WriteUndefined();
+                            break;
+                        case JsonToken.EndObject:
+                            writer.WriteEndObject();
+                            break;
+                        case JsonToken.EndArray:
+                            writer.WriteEndArray();
+                            break;
+                        case JsonToken.EndConstructor:
+                            writer.WriteEndConstructor();
+                            break;
+                        default:
+                            break;
+                    }
+                }
+
+            return sb.ToString();
         }
     }
 }
