+++ Orleans.Providers.MongoDB/StorageProviders/FileStorageProvider.cs
@@ -1,4 +1,6 @@
-﻿using System;
+﻿using MongoDB.Bson;
+using MongoDB.Bson.Serialization;
+using System;
 using System.IO;
 using System.Threading.Tasks;
 
@@ -82,7 +84,7 @@ namespace Orleans.Providers.MongoDB.StorageProviders
         /// <param name="collectionName">The type of the grain state object.</param>
         /// <param name="key">The grain id string.</param>
         /// <returns>Completion promise for this operation.</returns>
-        public async Task<string> Read(string collectionName, string key)
+        public async Task<BsonDocument> Read(string collectionName, string key)
         {
             var fileInfo = GetStorageFilePath(collectionName, key);
 
@@ -91,7 +93,8 @@ namespace Orleans.Providers.MongoDB.StorageProviders
 
             using (var stream = fileInfo.OpenText())
             {
-                return await stream.ReadToEndAsync();
+                var json = await stream.ReadToEndAsync();
+                return BsonSerializer.Deserialize<BsonDocument>(json);
             }
         }
 
@@ -102,13 +105,13 @@ namespace Orleans.Providers.MongoDB.StorageProviders
         /// <param name="key">The grain id string.</param>
         /// <param name="entityData">The grain state data to be stored./</param>
         /// <returns>Completion promise for this operation.</returns>
-        public async Task Write(string collectionName, string key, string entityData)
+        public async Task Write(string collectionName, string key, BsonDocument entityData)
         {
             var fileInfo = GetStorageFilePath(collectionName, key);
 
             using (var stream = new StreamWriter(fileInfo.Open(FileMode.Create, FileAccess.Write)))
             {
-                await stream.WriteAsync(entityData);
+                await stream.WriteAsync(entityData.ToJson());
             }
         }
